@@ -10,25 +10,28 @@
     "type": "library",
     "authors": [
         {
+            "name": "Aleksandar Milivojsa",
+            "email": "milivojsa96@gmail.com",
+            "role": "Developer"
+        },
+        {
             "name": "Mile Panić",
             "email": "mile.panic96@gmail.com",
             "role": "Developer"
         }
     ],
     "require": {
-        "php": "^8.0.2",
-        "illuminate/support": "^9.0",
+        "php": "^8.1.0",
         "spatie/laravel-activitylog": "^4.4",
-        "spatie/laravel-medialibrary": "^9.0.0",
+        "spatie/laravel-medialibrary": "^10.0.0",
         "spatie/laravel-permission": "^5.5",
         "spatie/laravel-sluggable": "^3.3",
         "spatie/laravel-view-models": "^1.3",
+        "symfony/dom-crawler": "^6.2",
         "vertex-it/blade-components": "^2.4"
     },
     "require-dev": {
-        "orchestra/testbench": "^6.0",
-        "phpunit/phpunit": "^9.0",
-        "vertex-it/php-cs": "^1.1"
+        "orchestra/testbench": "^8.0"
     },
     "autoload": {
         "psr-4": {
