@@ -14,10 +14,11 @@ class WorkTime implements CastsAttributes
      * @param  mixed  $value
      * @param  array  $attributes
      * @return mixed
+     * @throws \JsonException
      */
     public function get($model, string $key, $value, array $attributes): mixed
     {
-        return json_decode($value, true);
+        return json_decode($value, true, 512, JSON_THROW_ON_ERROR);
     }
 
     /**
@@ -28,13 +29,12 @@ class WorkTime implements CastsAttributes
      * @param  array  $value
      * @param  array  $attributes
      * @return string|bool
+     * @throws \JsonException
      */
     public function set($model, string $key, $value, array $attributes): string|bool
     {
-        return json_encode(
-            array_map(function ($day) {
-                return prepareMultipleInputData($day);
-            }, $value)
-        );
+        return json_encode(array_map(static function($day) {
+            return prepareMultipleInputData($day);
+        }, $value), JSON_THROW_ON_ERROR);
     }
 }
